Rename `assumed` to `builtins`
